+++ tests/Addressing.Tests/CountryCodeTests.cs
@@ -1,3 +1,5 @@
+using System.Reflection;
+
 namespace ISOCodex.Addressing.Tests;
 
 public class CountryCodeTests
@@ -17,6 +19,28 @@ public class CountryCodeTests
         Assert.Equal("US", CountryCode.US.Code);
         Assert.Equal("CA", CountryCode.CA.Code);
         Assert.Equal("ES", CountryCode.ES.Code);
+        Assert.Equal("FR", CountryCode.FR.Code);
+        Assert.Equal("IE", CountryCode.IE.Code);
+    }
+
+    [Fact]
+    public void NamedValues_ExistForEveryValidCountryCode()
+    {
+        var properties = typeof(CountryCode)
+            .GetProperties(BindingFlags.Public | BindingFlags.Static)
+            .Where(property => property.PropertyType == typeof(CountryCode))
+            .ToDictionary(property => property.Name, property => property);
+
+        foreach (var countryCode in CountryCode.All)
+        {
+            Assert.True(
+                properties.TryGetValue(countryCode.Code, out var property),
+                $"Expected CountryCode.{countryCode.Code} to be defined.");
+
+            var propertyValue = Assert.IsType<CountryCode>(property.GetValue(null));
+
+            Assert.Equal(countryCode, propertyValue);
+        }
     }
 
     [Fact]